requirements.txt
environs==9.3.3
requests==2.26.0
pip==21.2.4
attrs==21.2.0
pytz==2021.1
colorama==0.4.4
PyYAML==5.4.1
idna==3.2
multidict==5.1.0
Babel==2.9.1
mysql-connector-python==8.0.26
numpy==1.21.2
setuptools==58.0.4
regex==2021.8.28
torch==1.9.1
six==1.16.0
tqdm==4.62.3
urllib3==1.26.6
click==8.0.1
joblib==1.0.1
protobuf==3.18.0
aiohttp==3.7.4.post0
certifi==2021.5.30
aiogram==2.15
yarl==1.6.3
chardet==4.0.0
marshmallow==3.13.0
python-dotenv==0.19.0
pyparsing==2.4.7
sacremoses==0.0.45
tokenizers==0.10.3
filelock==3.0.12
transformers==4.10.2
packaging==21.0